ci: add github actions and integrate semantic-release
Relevant to 4324 in the React-Bootstrap repository.

Adds github actions, with a workflow for running tests and a
workflow for release.

For the release workflow, there is an integration with
semantic-release. We would need to need to setup two github
secrets for semantic to work properly: `NPM_TOKEN` and `GH_TOKEN`.