ci(actions): fix for build not triggering on forked pull request
Currently, the build is only being triggered whenever there is
a push to the repository.

This does not lead to a build being triggered whenever there is a
push to a branch on a forked repo.

This fix allows it to trigger on every push *or* every pull
request.

We'll probably need to also fix the configuration on the
React-Bootstrap repo.