Drop warning
I tend to remove warning because of confusing api. It's hard to say by
which condition warning is shown. This package also requires babel
plugin setup which is forgotted here. IMO it's much easier to just write
console statement and choose either console.warn or console.error than
dealing with this commonjs-only package.